Client-server
Started the implementation of classes:
- Server
- ClentThread
- Client

The Server starts threads, whence each of them communicates with exactly
one connected client, up to 4 connections.
The ClientThread represents one thread of the Server.
The Client communicates with the server, send him the player's
moves/actions, interpretes them into commands, which will be executed by
the Server and the information about the new state of the game field
will be returned to all connected clients.